+++ CHANGELOG.md
@@ -2,5 +2,10 @@
 
 All notable changes to the "vscode-translator-voice" extension will be documented in this file.
 
+## 0.2.0
+- Properties values in the extension app are updated with new configuration values without app restart - [Issue #1](https://github.com/yokawasa/vscode-translator-voice/issues/1)
+- Fixup bug: Toggled Voice state isn't persisted - [Issue #2](https://github.com/yokawasa/vscode-translator-voice/issues/2)
+- Changed StatusBarItem Label => `Voice [enabled/disabled]`
+
 ## 0.1.0
 - Initial release (alpha release)